MAINT/TEST: move tests and description of tests in separate file
