+++ statsmodels/distributions/copula/tests/test_copula.py
@@ -6,16 +6,20 @@ Author: Josef Perktold
 License: BSD-3
 
 """
-
 import numpy as np
-from numpy.testing import assert_allclose
+from numpy.testing import assert_allclose, assert_array_almost_equal
 from scipy import stats
 import pytest
+
+from statsmodels.distributions.copula.elliptical import GaussianCopula, \
+    StudentCopula
+from statsmodels.distributions.copula.other_copulas import IndependentCopula
 
 from statsmodels.tools.numdiff import approx_fprime_cs, approx_hess
 from statsmodels.distributions.copula.copulas import CopulaDistribution
 from statsmodels.distributions.copula.archimedean import (
-    ArchimedeanCopula, FrankCopula)
+    ArchimedeanCopula, FrankCopulaGeneratorBased, ClaytonCopula, FrankCopula,
+    GumbelCopula)
 from statsmodels.distributions.copula.extreme_value import (
     ExtremeValueCopula, copula_bv_ev)
 import statsmodels.distributions.copula.transforms as tra
@@ -239,13 +243,13 @@ def test_gev_genextreme(case):
     assert_allclose(pdfd, res2, rtol=1e-13)
 
 
-class TestFrank(object):
+class TestFrank:
 
     def test_basic(self):
         case = [tra.TransfFrank, 0.5, 0.9, (2,), 0.4710805107852225,
                 0.9257812360337806]
         cop_tr, v1, v2, args, cdf2, pdf2 = case
-        cop = FrankCopula()
+        cop = FrankCopulaGeneratorBased()
 
         pdf1 = cop.pdf([v1, v2], args=args)
         assert_allclose(pdf1, pdf2, rtol=1e-13)
@@ -262,3 +266,114 @@ class TestFrank(object):
         cdfc = cop.cdfcond_2g1(u, args=args)
         ppfc = cop.ppfcond_2g1(cdfc, [0.6], args=args)
         assert_allclose(ppfc, u[1], rtol=1e-13)
+
+
+# The reference results are coming from the R package Copula.
+# See ``copula_r_tests.rst`` for more details.
+
+
+class CopulaTests:
+    """Generic tests for copula."""
+    copula = None
+    dim = None
+    u = np.array([[0.33706249, 0.6075078],
+                  [0.62232507, 0.06241089],
+                  [0.2001457, 0.54027684],
+                  [0.77166391, 0.40610225],
+                  [0.98534253, 0.99212789],
+                  [0.72755898, 0.25913165],
+                  [0.05943888, 0.61044613],
+                  [0.0962475, 0.67585563],
+                  [0.35496733, 0.79584436],
+                  [0.44513594, 0.23050014]])
+    pdf_u = None
+    cdf_u = None
+
+    def test_visualization(self):
+        sample = self.copula.random(10000)
+        assert sample.shape == (10000, 2)
+        # h = sns.jointplot(sample[:, 0], sample[:, 1], kind='hex')
+        # h.set_axis_labels('X1', 'X2', fontsize=16)
+
+    def test_uniform_marginals(self):
+        sample = self.copula.random(10000)
+        assert_array_almost_equal(
+            np.mean(sample, axis=0), np.repeat(0.5, self.dim), decimal=2
+        )
+        assert_array_almost_equal(
+            np.percentile(sample, 25, axis=0), np.repeat(0.25, self.dim),
+            decimal=2
+        )
+        assert_array_almost_equal(
+            np.percentile(sample, 75, axis=0), np.repeat(0.75, self.dim),
+            decimal=2
+        )
+
+    def test_pdf(self):
+        pdf_u_test = self.copula.pdf(self.u)
+        assert_array_almost_equal(self.pdf_u, pdf_u_test)
+
+    def test_cdf(self):
+        cdf_u_test = self.copula.cdf(self.u)
+        assert_array_almost_equal(self.cdf_u, cdf_u_test)
+
+    def test_validate_params(self):
+        pass
+
+    def test_random_follow_pdf(self):
+        pass
+
+
+class TestIndependentCopula(CopulaTests):
+    copula = IndependentCopula()
+    dim = 2
+    pdf_u = np.ones((10, 1))
+    cdf_u = np.prod(CopulaTests.u, axis=1)
+
+
+class TestGaussianCopula(CopulaTests):
+    copula = GaussianCopula(cov=[[1., 0.8], [0.8, 1.]])
+    dim = 2
+    pdf_u = [1.03308741, 0.06507279, 0.72896012, 0.65389439, 16.45012399,
+             0.34813218, 0.06768115, 0.08168840, 0.40521741, 1.26723470]
+    cdf_u = [0.31906854, 0.06230196, 0.19284669, 0.39952707, 0.98144792,
+             0.25677003, 0.05932818, 0.09605404, 0.35211017, 0.20885480]
+
+
+class TestStudentCopula(CopulaTests):
+    copula = StudentCopula(cov=[[1., 0.8], [0.8, 1.]], df=2)
+    dim = 2
+    pdf_u = [0.8303065, 0.1359839, 0.5157746, 0.4776421, 26.2173959,
+             0.3070661, 0.1349173, 0.1597064, 0.3303230, 1.0482301]
+    cdf_u = [0.31140349, 0.05942746, 0.18548601, 0.39143974, 0.98347259,
+             0.24894028, 0.05653947, 0.09210693, 0.34447385, 0.20429882]
+
+    def test_cdf(self, *args):
+        pytest.skip("Not implemented.")
+
+
+class TestClaytonCopula(CopulaTests):
+    copula = ClaytonCopula(theta=1.2)
+    dim = 2
+    pdf_u = [1.0119836, 0.2072728, 0.8148839, 0.9481976, 2.1419659,
+             0.6828507, 0.2040454, 0.2838497, 0.8197787, 1.1096360]
+    cdf_u = [0.28520375, 0.06101690, 0.17703377, 0.36848218, 0.97772088,
+             0.24082057, 0.05811908, 0.09343934, 0.33012582, 0.18738753]
+
+
+class TestFrankCopula(CopulaTests):
+    copula = FrankCopula(theta=3)
+    dim = 2
+    pdf_u = [0.9646599, 0.5627195, 0.8941964, 0.8364614, 2.9570945,
+             0.6665601, 0.5779906, 0.5241333, 0.7156741, 1.1074024]
+    cdf_u = [0.27467496, 0.05492539, 0.15995939, 0.36750702, 0.97782283,
+             0.23412757, 0.05196265, 0.08676979, 0.32803721, 0.16320730]
+
+
+class TestGumbelCopula(CopulaTests):
+    copula = GumbelCopula(theta=1.5)
+    dim = 2
+    pdf_u = [1.0391696, 0.6539579, 0.9878446, 0.8679504, 16.6030932,
+             0.7542073, 0.6668307, 0.6275887, 0.7477991, 1.1564864]
+    cdf_u = [0.27194634, 0.05484380, 0.15668190, 0.37098420, 0.98176346,
+             0.23422865, 0.05188260, 0.08659615, 0.33086960, 0.15803914]
